Fix: Region tile loading for cache 209

+++ src/main/kotlin/cache/definitions/RegionDefinition.kt
@@ -82,10 +82,10 @@ class RegionDefinition(
         var height: Int = 0,
         var attrOpcode: Int = 0,
         var settings: Byte = 0,
-        var overlayId: Byte = 0,
+        var overlayId: Short = 0,
         var overlayPath: Byte = 0,
         var overlayRotation: Byte = 0,
-        var underlayId: Byte = 0
+        var underlayId: Short = 0
     )
 
     companion object {
